requirements.txt
asgiref==3.3.1
Django==3.1.5
django-ckeditor==6.0.0
django-ckeditor-widget==0.1
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-modeltranslation==0.16.1
django-translations==1.2.0
mysqlclient==2.0.3
Pillow==8.1.0
python-gettext==4.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
